+++ config.php
@@ -4,6 +4,7 @@
 
 use humhub\commands\CronController;
 use humhub\components\Controller;
+use humhub\components\gates\GateManager;
 use humhub\modules\content\widgets\richtext\ProsemirrorRichText;
 use humhub\modules\user\models\forms\Registration;
 use humhub\modules\user\widgets\AccountSettingsMenu;
@@ -25,6 +26,7 @@ return [
         ['class' => LayoutAddons::class, 'event' => LayoutAddons::EVENT_INIT, 'callback' => ['humhub\modules\legal\Events', 'onLayoutAddonInit']],
         ['class' => Registration::class, 'event' => Registration::EVENT_AFTER_INIT, 'callback' => ['humhub\modules\legal\Events', 'onRegistrationFormInit']],
         ['class' => Registration::class, 'event' => Registration::EVENT_AFTER_REGISTRATION, 'callback' => ['humhub\modules\legal\Events', 'onRegistrationAfterRegistration']],
+        ['class' => GateManager::class, 'event' => GateManager::EVENT_INIT_GATES, 'callback' => ['humhub\modules\legal\Events', 'onGateInit']],
         ['class' => Controller::class, 'event' => Controller::EVENT_BEFORE_ACTION, 'callback' => ['humhub\modules\legal\Events', 'onBeforeControllerAction']],
         ['class' => ProsemirrorRichText::class, 'event' => ProsemirrorRichText::EVENT_AFTER_RUN, 'callback' => ['humhub\modules\legal\Events', 'onAfterRunRichText']],
         ['class' => AccountSettingsMenu::class, 'event' => AccountSettingsMenu::EVENT_INIT, 'callback' => ['humhub\modules\legal\Events', 'onAccountSettingsMenuInit']],